chore: move rts caller interface to external module (#39103)
## Description
> [!IMPORTANT]  
> Move RTSCaller as an interface to external module.


Fixes #`Issue Number`  
_or_  
Fixes `Issue URL`
> [!WARNING]  
> _If no issue exists, please create an issue first, and check with the
maintainers if the issue is valid._

## Automation

/ok-to-test tags="@tag.Sanity"

### :mag: Cypress test results
