+++ app/server/appsmith-interfaces/src/main/java/com/appsmith/external/services/RTSCaller.java
@@ -1,0 +1,5 @@
+package com.appsmith.external.services;
+
+import com.appsmith.external.services.ce.RTSCallerCE;
+
+public interface RTSCaller extends RTSCallerCE {}
